updated #525

+++ tests/golos.posting_test_api.hpp
@@ -50,7 +50,7 @@ struct golos_posting_api: base_contract_api {
         std::string language = "languagemssg",
         std::vector<tags> tags = {{"tag"}},
         std::string json_metadata = "jsonmetadata",
-        int64_t curators_prcnt = 7100 
+        uint16_t curators_prcnt = 7100 
     ) {
         return push(N(createmssg), message_id.author, args()
             ("message_id", message_id)
@@ -112,7 +112,7 @@ struct golos_posting_api: base_contract_api {
             ("message_id", message_id)
         );
     }
-    action_result set_curators_prcnt(mssgid message_id, int64_t curators_prcnt) {
+    action_result set_curators_prcnt(mssgid message_id, uint16_t curators_prcnt) {
         return push(N(setcurprcnt), message_id.author, args()
             ("message_id", message_id)
             ("curators_prcnt", curators_prcnt)
@@ -166,7 +166,7 @@ struct golos_posting_api: base_contract_api {
         return string("['st_referral_acc', {'value':'") + name{referral_acc}.to_string() + "'}]";
     }   
     
-    string get_str_curators_prcnt(int64_t min_curators_prcnt, int64_t max_curators_prcnt) {
+    string get_str_curators_prcnt(uint16_t min_curators_prcnt, uint16_t max_curators_prcnt) {
         return string("['st_curators_prcnt', {'min_curators_prcnt':'") + std::to_string(min_curators_prcnt) + "','max_curators_prcnt':'" + std::to_string(max_curators_prcnt) + "'}]";
     }
 
@@ -211,8 +211,8 @@ struct golos_posting_api: base_contract_api {
     const uint32_t upvote_lockout = 15;
     const uint8_t max_beneficiaries = 64;
     const uint16_t max_comment_depth = 127;
-    const int64_t min_curators_prcnt = 0;
-    const int64_t max_curators_prcnt = 9000;
+    const uint16_t min_curators_prcnt = 0;
+    const uint16_t max_curators_prcnt = 9000;
 };
